+++ packages/core/src/decoder/index.js
@@ -1,0 +1,1 @@
+export { decodeAction } from './actionDecoder.js';